# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Weather API",
    "version": "1.0.0",
    "description": "API for retrieving and storing weather data"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
    "/weather": {
      "post": {
        "summary": "Create new weather data",
        "description": "Add weather data for a city to the database",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "city": {
                    "type": "string"
                  },
                  "temperature": {
                    "type": "number"
                  },
                  "description": {
                    "type": "string"
                  },
                  "humidity": {
                    "type": "number"
                  },
                  "pressure": {
                    "type": "number"
                  },
                  "windSpeed": {
                    "type": "number"
                  },
                  "windDirection": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Weather data created successfully"
          },
          "500": {
            "description": "Error creating weather data"
          }
        }
      }
    },
    "/weather/data/{city}": {
      "get": {
        "summary": "Get weather data by city",
        "description": "Retrieve weather data for a specific city",
        "parameters": [
          {
            "in": "path",
            "name": "city",
            "required": true,
            "description": "The city name for which you want to fetch the weather data.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved weather data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "city": {
                        "type": "string"
                      },
                      "temperature": {
                        "type": "number"
                      },
                      "description": {
                        "type": "string"
                      },
                      "humidity": {
                        "type": "number"
                      },
                      "pressure": {
                        "type": "number"
                      },
                      "windSpeed": {
                        "type": "number"
                      },
                      "windDirection": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No weather data found for the specified city"
          }
        }
      }
    },
    "/weather/update/{city}": {
      "put": {
        "summary": "Update weather data for a city",
        "description": "Update existing weather data for a specific city",
        "parameters": [
          {
            "in": "path",
            "name": "city",
            "required": true,
            "description": "The city name for which you want to update the weather data.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "temperature": {
                    "type": "number"
                  },
                  "description": {
                    "type": "string"
                  },
                  "humidity": {
                    "type": "number"
                  },
                  "pressure": {
                    "type": "number"
                  },
                  "windSpeed": {
                    "type": "number"
                  },
                  "windDirection": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated weather data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "city": {
                      "type": "string"
                    },
                    "temperature": {
                      "type": "number"
                    },
                    "description": {
                      "type": "string"
                    },
                    "humidity": {
                      "type": "number"
                    },
                    "pressure": {
                      "type": "number"
                    },
                    "windSpeed": {
                      "type": "number"
                    },
                    "windDirection": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Weather data not found for the specified city"
          },
          "500": {
            "description": "Error updating weather data"
          }
        }
      }
    },
    "/weather/delete/{city}": {
      "delete": {
        "summary": "Delete weather data for a city",
        "description": "Delete the weather data of a specific city",
        "parameters": [
          {
            "in": "path",
            "name": "city",
            "required": true,
            "description": "The city name for which you want to delete the weather data.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted weather data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No weather data found for the specified city to delete"
          },
          "500": {
            "description": "Error deleting weather data"
          }
        }
      }
    },
    "/api/location": {
      "post": {
        "summary": "Add a new location",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Location"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Location created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          },
          "500": {
            "description": "Failed to save location"
          }
        }
      },
      "get": {
        "summary": "Get all locations",
        "responses": {
          "200": {
            "description": "List of all locations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Location"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch locations"
          }
        }
      }
    },
    "/api/location/{city}": {
      "get": {
        "summary": "Get a specific location by city",
        "parameters": [
          {
            "in": "path",
            "name": "city",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Name of the city"
          }
        ],
        "responses": {
          "200": {
            "description": "Location details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          },
          "404": {
            "description": "Location not found"
          },
          "500": {
            "description": "Failed to fetch location"
          }
        }
      },
      "put": {
        "summary": "Update a location by city",
        "parameters": [
          {
            "in": "path",
            "name": "city",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Name of the city"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Location"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Location updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          },
          "404": {
            "description": "Location not found"
          },
          "500": {
            "description": "Failed to update location"
          }
        }
      },
      "delete": {
        "summary": "Delete a location by city",
        "parameters": [
          {
            "in": "path",
            "name": "city",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Name of the city"
          }
        ],
        "responses": {
          "200": {
            "description": "Location deleted successfully"
          },
          "404": {
            "description": "Location not found"
          },
          "500": {
            "description": "Failed to delete location"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Location": {
        "type": "object",
        "properties": {
          "city": {
            "type": "string",
            "description": "The name of the city"
          },
          "country": {
            "type": "string",
            "description": "The country where the location is"
          },
          "latitude": {
            "type": "number",
            "description": "The latitude of the location"
          },
          "longitude": {
            "type": "number",
            "description": "The longitude of the location"
          }
        }
      }
    }
  },
  "tags": []
}